Remove most of the rest of posts
