+++ tests/Feature/Tags/ListTagsTest.php
@@ -1,9 +1,9 @@
 <?php declare(strict_types=1);
 
-use Siteman\Cms\Models\Post;
+use Siteman\Cms\Models\Page;
 
 it('can list tags with published posts count', function () {
-    $posts = Post::factory()->count(2)->published()->withTags(['foo', 'bar'])->create();
+    $posts = Page::factory()->count(2)->published()->withTags(['foo', 'bar'])->create();
 
     $this->get('/tags/foo')
         ->assertOk()
